chore: 패키지명 및 구조 변경
