--- src/main/java/subway/application/LineRepository.java
@@ -1,12 +1,0 @@
-package subway.application;
-
-import subway.domain.Line;
-
-public interface LineRepository {
-
-    Line findById(Long linePropertyId);
-
-    void removeSections(Long lineId);
-
-    void save(Line line);
-}
