--- src/main/java/subway/application/StationService.java
@@ -1,46 +1,0 @@
-package subway.application;
-
-import org.springframework.stereotype.Service;
-import org.springframework.transaction.annotation.Transactional;
-import subway.domain.Station;
-import subway.dto.StationRequest;
-import subway.dto.StationResponse;
-
-import java.util.List;
-import java.util.stream.Collectors;
-
-@Service
-@Transactional
-public class StationService {
-
-    private final StationRepository stationRepository;
-
-    public StationService(StationRepository stationRepository) {
-        this.stationRepository = stationRepository;
-    }
-
-    public StationResponse saveStation(StationRequest stationRequest) {
-        Station station = stationRepository.insert(new Station(stationRequest.getName()));
-        return StationResponse.of(station);
-    }
-
-    public StationResponse findStationResponseById(Long id) {
-        return StationResponse.of(stationRepository.findById(id));
-    }
-
-    public List<StationResponse> findAllStationResponses() {
-        List<Station> stations = stationRepository.findAll();
-
-        return stations.stream()
-                .map(StationResponse::of)
-                .collect(Collectors.toList());
-    }
-
-    public void updateStation(Long id, StationRequest stationRequest) {
-        stationRepository.update(new Station(id, stationRequest.getName()));
-    }
-
-    public void deleteStationById(Long id) {
-        stationRepository.deleteById(id);
-    }
-}
