+++ src/main/java/subway/application/repository/LinePropertyRepository.java
@@ -1,0 +1,18 @@
+package subway.application.repository;
+
+import subway.application.domain.LineProperty;
+
+import java.util.List;
+
+public interface LinePropertyRepository {
+
+    LineProperty insert(LineProperty lineProperty);
+
+    List<LineProperty> findAll();
+
+    LineProperty findById(Long id);
+
+    void update(LineProperty lineProperty);
+
+    void deleteById(Long id);
+}
