--- src/main/java/subway/domain/LineProperty.java
@@ -1,41 +1,0 @@
-package subway.domain;
-
-import java.util.Objects;
-
-public class LineProperty {
-
-    private final Long id;
-    private final String name;
-    private final String color;
-
-    public LineProperty(Long id, String name, String color) {
-        this.id = id;
-        this.name = name;
-        this.color = color;
-    }
-
-    public Long getId() {
-        return id;
-    }
-
-    public String getName() {
-        return name;
-    }
-
-    public String getColor() {
-        return color;
-    }
-
-    @Override
-    public boolean equals(Object o) {
-        if (this == o) return true;
-        if (o == null || getClass() != o.getClass()) return false;
-        LineProperty that = (LineProperty) o;
-        return id.equals(that.id);
-    }
-
-    @Override
-    public int hashCode() {
-        return Objects.hash(id);
-    }
-}
