+++ src/main/java/subway/persistence/H2StationRepository.java
@@ -1,0 +1,60 @@
+package subway.persistence;
+
+import org.springframework.jdbc.core.JdbcTemplate;
+import org.springframework.jdbc.core.RowMapper;
+import org.springframework.jdbc.core.namedparam.BeanPropertySqlParameterSource;
+import org.springframework.jdbc.core.namedparam.SqlParameterSource;
+import org.springframework.jdbc.core.simple.SimpleJdbcInsert;
+import org.springframework.stereotype.Repository;
+import subway.application.repository.StationRepository;
+import subway.application.domain.Station;
+
+import javax.sql.DataSource;
+import java.util.List;
+
+@Repository
+public class H2StationRepository implements StationRepository {
+
+    private final JdbcTemplate jdbcTemplate;
+    private final SimpleJdbcInsert insertAction;
+
+    private RowMapper<Station> rowMapper = (rs, rowNum) ->
+            new Station(
+                    rs.getLong("id"),
+                    rs.getString("name")
+            );
+
+
+    public H2StationRepository(JdbcTemplate jdbcTemplate, DataSource dataSource) {
+        this.jdbcTemplate = jdbcTemplate;
+        this.insertAction = new SimpleJdbcInsert(dataSource)
+                .withTableName("station")
+                .usingGeneratedKeyColumns("id");
+    }
+
+    public Station insert(Station station) {
+        SqlParameterSource params = new BeanPropertySqlParameterSource(station);
+        Long id = insertAction.executeAndReturnKey(params).longValue();
+        return new Station(id, station.getName());
+    }
+
+    public List<Station> findAll() {
+        String sql = "select * from STATION";
+        return jdbcTemplate.query(sql, rowMapper);
+    }
+
+    public Station findById(Long id) {
+        String sql = "select * from STATION where id = ?";
+        return jdbcTemplate.queryForObject(sql, rowMapper, id);
+    }
+
+    public void update(Station newStation) {
+        String sql = "update STATION set name = ? where id = ?";
+        jdbcTemplate.update(sql, new Object[]{newStation.getName(), newStation.getId()});
+    }
+
+    public void deleteById(Long id) {
+        String sql = "delete from STATION where id = ?";
+        jdbcTemplate.update(sql, id);
+    }
+}
