+++ src/main/java/subway/presentation/controller/LinePropertyController.java
@@ -1,0 +1,55 @@
+package subway.presentation.controller;
+
+import org.springframework.http.ResponseEntity;
+import org.springframework.web.bind.annotation.*;
+import subway.application.service.LinePropertyService;
+import subway.presentation.dto.LineRequest;
+import subway.presentation.dto.LineResponse;
+
+import java.net.URI;
+import java.sql.SQLException;
+import java.util.List;
+
+@RestController
+@RequestMapping("/lines")
+public class LinePropertyController {
+
+    private final LinePropertyService linePropertyService;
+
+    public LinePropertyController(LinePropertyService linePropertyService) {
+        this.linePropertyService = linePropertyService;
+    }
+
+    @PostMapping
+    public ResponseEntity<LineResponse> createLine(@RequestBody LineRequest lineRequest) {
+        LineResponse line = linePropertyService.saveLine(lineRequest);
+        return ResponseEntity.created(URI.create("/lines/" + line.getId())).body(line);
+    }
+
+    @GetMapping
+    public ResponseEntity<List<LineResponse>> findAllLines() {
+        return ResponseEntity.ok(linePropertyService.findLineResponses());
+    }
+
+    @GetMapping("/{id}")
+    public ResponseEntity<LineResponse> findLineById(@PathVariable Long id) {
+        return ResponseEntity.ok(linePropertyService.findLineResponseById(id));
+    }
+
+    @PutMapping("/{id}")
+    public ResponseEntity<Void> updateLine(@PathVariable Long id, @RequestBody LineRequest lineUpdateRequest) {
+        linePropertyService.updateLine(id, lineUpdateRequest);
+        return ResponseEntity.ok().build();
+    }
+
+    @DeleteMapping("/{id}")
+    public ResponseEntity<Void> deleteLine(@PathVariable Long id) {
+        linePropertyService.deleteLineById(id);
+        return ResponseEntity.noContent().build();
+    }
+
+    @ExceptionHandler(SQLException.class)
+    public ResponseEntity<Void> handleSQLException() {
+        return ResponseEntity.badRequest().build();
+    }
+}
