+++ src/main/java/subway/presentation/dto/StationResponse.java
@@ -1,0 +1,26 @@
+package subway.presentation.dto;
+
+import subway.application.domain.Station;
+
+public class StationResponse {
+
+    private Long id;
+    private String name;
+
+    public StationResponse(Long id, String name) {
+        this.id = id;
+        this.name = name;
+    }
+
+    public static StationResponse of(Station station) {
+        return new StationResponse(station.getId(), station.getName());
+    }
+
+    public Long getId() {
+        return id;
+    }
+
+    public String getName() {
+        return name;
+    }
+}
