+++ src/test/java/subway/domain/DistanceTest.java
@@ -3,6 +3,7 @@ package subway.domain;
 import org.junit.jupiter.api.DisplayName;
 import org.junit.jupiter.params.ParameterizedTest;
 import org.junit.jupiter.params.provider.ValueSource;
+import subway.application.domain.Distance;
 
 import static org.assertj.core.api.AssertionsForClassTypes.assertThatThrownBy;
