+++ src/test/java/subway/domain/LineTest.java
@@ -3,6 +3,11 @@ package subway.domain;
 import org.junit.jupiter.api.BeforeEach;
 import org.junit.jupiter.api.DisplayName;
 import org.junit.jupiter.api.Test;
+import subway.application.domain.Distance;
+import subway.application.domain.Line;
+import subway.application.domain.LineProperty;
+import subway.application.domain.Section;
+import subway.application.domain.Station;
 
 import java.util.ArrayList;
